requirements.txt
pygpt4all==1.1.0
chromadb==0.3.23
llama-cpp-python==0.1.38
urllib3==2.0.2
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
